refactor: api views to separate modules
